+++ www/modules/smime/language/ro.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  'To upload a new PKCS12 certificate you must enter your Group-Office password. The Group-Office password must be different than your PCSK12 certificate for security reasons. No password is also prohibited.' => 'To upload a new PKCS12 certificate you must enter your Safety Broker password. The Safety Broker password must be different than your PCSK12 certificate for security reasons. No password is also prohibited.',
+);